// packages/Morado/src/store/components/Product/types.ts
interface QuantitySelectorProps {
  availableQuantityInfo?: boolean;
}

interface CartButtonProps {
  text?: string;
  variant?: 'default' | 'slim' | 'simple';
}
type ICard = {
  image: ImageProps;
  product: Product;
  noNameTag?: boolean;
  maxCharts?: number;
  style?: string;
  cart?: CartButtonProps;
  quantitySelector?: QuantitySelectorProps;
};